pull before pr

+++ PaddleNLP/unarchived/deep_attention_matching_net/utils/reader.py
@@ -1,3 +1,16 @@
+#   Copyright (c) 2019 PaddlePaddle Authors. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 import six
 import numpy as np
 
@@ -188,11 +201,12 @@ def make_one_batch_input(data_batches, index):
         feeding dictionary
     """
 
-    turns = np.array(data_batches["turns"][index])
-    tt_turns_len = np.array(data_batches["tt_turns_len"][index])
-    every_turn_len = np.array(data_batches["every_turn_len"][index])
-    response = np.array(data_batches["response"][index])
-    response_len = np.array(data_batches["response_len"][index])
+    turns = np.array(data_batches["turns"][index]).astype('int64')
+    tt_turns_len = np.array(data_batches["tt_turns_len"][index]).astype('int64')
+    every_turn_len = np.array(data_batches["every_turn_len"][index]).astype(
+        'int64')
+    response = np.array(data_batches["response"][index]).astype('int64')
+    response_len = np.array(data_batches["response_len"][index]).astype('int64')
 
     batch_size = turns.shape[0]
     max_turn_num = turns.shape[1]
